refactor (armazenamento)
trocado “float” por “double” para maior precisão

+++ caractere.h
@@ -2,7 +2,7 @@
 #define CARACTERE_H
 
 
-float func_val(int x, float b);
+double func_val(int x, float b);
 void decodificador(char codigo[101], float b);
 
 #endif
